Changed all PositiveIntegerFields in all models which can be float to FloatField

+++ econom_game/cards/models.py
@@ -4,7 +4,7 @@ from django.db import models
 class Card(models.Model):
     card_number = models.CharField(max_length=25, unique=True)
     chip_number = models.CharField(max_length=25, unique=True)
-    money_amount = models.PositiveIntegerField()
+    money_amount = models.FloatField()
 
     def __str__(self):
         return 'card_%d' % self.id
